+++ e2e/pyproject.toml
@@ -18,6 +18,7 @@ requests = "^2.31.0"
 pyjwt = "^2.8.0"
 cryptography = "^42.0.3"
 lxml = "~4.9.0"
+debugpy = "^1.8.13"
 
 [build-system]
 requires = ["poetry-core"]
