+++ e2e/ssocli-bundle/install
@@ -1,0 +1,50 @@
+#!/bin/bash
+
+set -e
+
+usage_warning() {
+
+  echo "
+  Installs scripts to generate OAuth tokens from apigee single-sign-on (sso) service:
+
+  Default installation path for scripts is \"/usr/local/bin\" .
+  use -b <path_location> to override the default path.
+  e.g. ./install.sh -b /usr/local/bin
+
+  "
+  exit 1
+
+}
+
+while getopts ":b:" opt; do
+  case ${opt} in
+    b)
+      INSTALL_DIR="${OPTARG}"
+      ;;
+    \?)
+      echo "Invalid option: -${OPTARG}" >&2
+      usage_warning
+      ;;
+    :)
+      echo "Option -${OPTARG} requires an argument." >&2
+      usage_warning
+      ;;
+  esac
+done
+
+if [[ -z "${INSTALL_DIR}" ]]; then
+  INSTALL_DIR="/usr/local/bin"
+fi
+
+if [ ! -d "${INSTALL_DIR}" ]; then
+  echo "${INSTALL_DIR} directory does not exist!"
+  exit 1
+fi
+
+CUR_DIR=$(pwd)
+echo "Current working dir: ${CUR_DIR}"
+
+cp "${CUR_DIR}/acurl" "${INSTALL_DIR}"
+cp "${CUR_DIR}/get_token" "${INSTALL_DIR}"
+
+echo "Successfully installed scripts to location: ${INSTALL_DIR}"
